requirements.txt
albumentations==0.5.2
anyio==2.2.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
bleach==3.3.0
boto3==1.17.56
botocore==1.20.56
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.7.0
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
filelock==3.0.12
gevent==21.1.2
glob2==0.7
greenlet==1.0.0
idna==2.9
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.0.1
inotify-simple==1.2.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.2
jupyter-server==1.6.4
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
libarchive-c==2.9
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==2.5.0
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numpy==1.19.5
olefile==0.46
opencv-python==4.2.0.34
opencv-python-headless==4.5.1.48
packaging==20.9
pandas==1.1.1
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
pretrainedmodels==0.7.4
prometheus-client==0.10.1
prompt-toolkit==3.0.5
protobuf==3.15.8
psutil==5.7.0
ptyprocess==0.6.0
pycocotools==2.0.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
requests==2.23.0
retrying==1.3.3
ruamel-yaml==0.15.87
s3transfer==0.4.2
sagemaker-training==3.9.1
scikit-image==0.18.1
scipy==1.6.2
seaborn==0.11.1
segmentation-models-pytorch==0.1.3
Send2Trash==1.5.0
Shapely==1.7.1
six==1.14.0
sniffio==1.2.0
soupsieve==2.0.1
terminado==0.9.4
testpath==0.4.4
tifffile==2021.4.8
timm==0.3.2
tomlkit==0.7.0
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
